+++ projects/sample/types/stores/GameStateActions.ts
@@ -1,4 +1,4 @@
-import type { GameState } from './GameState';
+import type { GameState } from '@/generate/types';
 
 export interface GameStateActions {
   resetGame(this: GameState): void;
